refactor(docker): extract ContainerHistoryPage data logic into custom hook
Queries, debouncing, and metric selection moved to useContainerHistoryData.
ContainerHistoryPage.tsx is now a thin rendering layer.
